+++ src/lib.rs
@@ -8,6 +8,6 @@ pub mod asciigraph;
 
 pub use crate::color::AnsiColor;
 pub use crate::options::{
-    CharSet, Config, ZeroLine, Threshold, DEFAULT_CHAR_SET, create_char_set
+    CharSet, Config, ZeroLine, Threshold, StatAnnotations, DEFAULT_CHAR_SET, create_char_set
 };
 pub use crate::asciigraph::{plot, plot_many};